requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Limiter==4.0.0
python-dotenv==1.0.0
python-telegram-bot==20.7
requests==2.31.0
SQLAlchemy==2.0.23
pybind11==2.11.1
APScheduler==3.10.4
gunicorn==21.2.0
Werkzeug==3.0.1
psycopg2-binary==2.9.11
